Refactoring the project as follow: - refactoring project to use Python 3.6 version (ie. print functions); - create file neuralnetwork.py, where is implemented it's features;
