+++ README.md
@@ -4,4 +4,9 @@ A neural network written in Python, consisting of a single neuron that uses back
 ### Install Dependencies
 1. First, install PIP package on your pc. (https://pip.pypa.io/en/stable/installing/)
 2. Install project dependencies: pip install -r requirements.txt
-3. Have fun!
+3. How to run? 
+    * python3.6 **main.py**
+     
+    * python3.6 **short-version.py**
+    
+Ps: This project use Python 3.6 version only.
