+++ neuralnetwork.py
@@ -1,0 +1,51 @@
+from numpy import exp, random, dot
+
+
+class NeuralNetwork:
+    def __init__(self):
+        # Seed the random number generator, so it generates the same numbers
+        # every time the program runs.
+        random.seed(1)
+
+        # We model a single neuron, with 3 input connections and 1 output connection.
+        # We assign random weights to a 3 x 1 matrix, with values in the range -1 to 1
+        # and mean 0.
+        self.synaptic_weights = 2 * random.random((3, 1)) - 1
+
+    # The Sigmoid function, which describes an S shaped curve.
+    # We pass the weighted sum of the inputs through this function to
+    # normalise them between 0 and 1.
+    @staticmethod
+    def __sigmoid(x):
+        return 1 / (1 + exp(-x))
+
+    # The derivative of the Sigmoid function.
+    # This is the gradient of the Sigmoid curve.
+    # It indicates how confident we are about the existing weight.
+    @staticmethod
+    def __sigmoid_derivative(x):
+        return x * (1 - x)
+
+    # We train the neural network through a process of trial and error.
+    # Adjusting the synaptic weights each time.
+    def train(self, train_set_inputs, train_set_outputs, number_of_training_iterations):
+        for iteration in range(number_of_training_iterations):
+            # Pass the training set through our neural network (a single neuron).
+            output = self.think(train_set_inputs)
+
+            # Calculate the error (The difference between the desired output
+            # and the predicted output).
+            error = train_set_outputs - output
+
+            # Multiply the error by the input and again by the gradient of the Sigmoid curve.
+            # This means less confident weights are adjusted more.
+            # This means inputs, which are zero, do not cause changes to the weights.
+            adjustment = dot(train_set_inputs.T, error * self.__sigmoid_derivative(output))
+
+            # Adjust the weights.
+            self.synaptic_weights += adjustment
+
+    # The neural network thinks.
+    def think(self, inputs):
+        # Pass inputs through our neural network (our single neuron).
+        return self.__sigmoid(dot(inputs, self.synaptic_weights))
