+++ short_version.py
@@ -3,7 +3,8 @@ training_set_inputs = array([[0, 0, 1], [1, 1, 1], [1, 0, 1], [0, 1, 1]])
 training_set_outputs = array([[0, 1, 1, 0]]).T
 random.seed(1)
 synaptic_weights = 2 * random.random((3, 1)) - 1
-for iteration in xrange(10000):
+
+for iteration in range(10000):
     output = 1 / (1 + exp(-(dot(training_set_inputs, synaptic_weights))))
     synaptic_weights += dot(training_set_inputs.T, (training_set_outputs - output) * output * (1 - output))
-print 1 / (1 + exp(-(dot(array([1, 0, 0]), synaptic_weights))))
+print(1 / (1 + exp(-(dot(array([1, 0, 0]), synaptic_weights)))))
